+++ src/mako/application/services/LoggerService.php
@@ -8,9 +8,11 @@
 namespace mako\application\services;
 
 use mako\application\Application;
+use mako\gatekeeper\Authentication;
+use mako\logger\Logger;
 use Monolog\Formatter\LineFormatter;
+use Monolog\Handler\HandlerInterface;
 use Monolog\Handler\StreamHandler;
-use Monolog\Logger;
 use Psr\Log\LoggerInterface;
 
 use function date;
@@ -23,19 +25,50 @@ use function date;
 class LoggerService extends Service
 {
 	/**
+	 * Get global logger context.
+	 *
+	 * @return array
+	 */
+	protected function getContext(): array
+	{
+		if($this->container->has(Authentication::class) && ($user = $this->container->get(Authentication::class)->getUser()) !== null)
+		{
+			return ['user_id' => $user->getId()];
+		}
+
+		return [];
+	}
+
+	/**
+	 * Returns the default handler.
+	 *
+	 * @return \Monolog\Handler\HandlerInterface
+	 */
+	protected function getHandler(): HandlerInterface
+	{
+		$handler = new StreamHandler($this->container->get(Application::class)->getPath() . '/storage/logs/' . date('Y-m-d') . '.mako');
+
+		$formatter = new LineFormatter(null, null, true, true);
+
+		$formatter->includeStacktraces();
+
+		$handler->setFormatter($formatter);
+
+		return $handler;
+	}
+
+	/**
 	 * {@inheritdoc}
 	 */
 	public function register(): void
 	{
-		$this->container->registerSingleton([LoggerInterface::class, 'logger'], function($container)
+		$this->container->registerSingleton([LoggerInterface::class, 'logger'], function()
 		{
 			$logger = new Logger('mako');
 
-			$handler = new StreamHandler($container->get(Application::class)->getPath() . '/storage/logs/' . date('Y-m-d') . '.mako', Logger::DEBUG);
+			$logger->setContext($this->getContext());
 
-			$handler->setFormatter(new LineFormatter(null, null, true, true));
-
-			$logger->pushHandler($handler);
+			$logger->pushHandler($this->getHandler());
 
 			return $logger;
 		});